Get cpp code working + add Xcode proj
Unclear why pq tests were segfaulting (only with Bazel) before. Everything
runs now, but needs to be cleaned up; this commit is mostly so that I have
something working to compare to as I clean the code and try to avoid
mysteriously recreating the error.

Also, note that the command under which the Bazel version works is:

>>> .../bolt/cpp$ CC=clang blaze run :main -- '[pq]'